bug fix , settings module

+++ src/Components/Navbar.jsx
@@ -23,7 +23,7 @@ const Navbar = ({ openModal }) => {
     // Get current date
     useEffect(() => {
         const today = new Date();
-        const options = { weekday: "long", month: "long", day: "numeric" };
+        const options = { weekday: "short", month: "short", day: "numeric" };
         setDate(today.toLocaleDateString("en-US", options));
     }, []);
